+++ rvs/resources/__init__.py
@@ -1,0 +1,1 @@
+"""Static trust resources bundled with every rvs distribution."""